requirements.txt
streamlit
langchain
langchain_pinecone
pinecone-client
sentence-transformers
numpy
PyPDF2
groq